+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- chevreと重複するタスク名をrename
+
 ### Deprecated
 
 ### Removed